comment out unused commands, rename files, move files
