+++ workspace sir vivor/commands/text.js
@@ -164,6 +164,5 @@ module.exports = {
 	
 	saw: '==workshop',
 	workshop: '<<survivorworkshop>>',
-	
 }
